Introduced OQ_CHECK_INPUT [skip hazardlib]
